Code polishing
