+++ NorthwindCRUD/QueryBuilder/Model/FilterType.cs
@@ -1,0 +1,7 @@
+﻿namespace QueryBuilder;
+
+public enum FilterType
+{
+    And = 0,
+    Or = 1,
+}
