+++ NorthwindCRUD/QueryBuilder/Model/QueryFilterCondition.cs
@@ -1,0 +1,10 @@
+﻿namespace QueryBuilder;
+
+public class QueryFilterCondition
+{
+    public string Name { get; set; }
+
+    public bool IsUnary { get; set; }
+
+    public string IconName { get; set; }
+}
